requirements.txt
Flask
google-genai
opencv-python
pyaudio
pillow
mss
